Move request to http package.
Add an alias for backwards compatibility.

Refs #9325
